day 2/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
jinja2
